+++ x/posts/internal/types/post.go
@@ -104,17 +104,17 @@ func (ids PostIDs) String() string {
 
 // Post is a struct of a post
 type Post struct {
-	PostID         PostID         `json:"id"`                      // Unique id
-	ParentID       PostID         `json:"parent_id"`               // Post of which this one is a comment
-	Message        string         `json:"message"`                 // Message contained inside the post
-	Created        time.Time      `json:"created"`                 // RFC3339 date at which the post has been created
-	LastEdited     time.Time      `json:"last_edited"`             // RFC3339 date at which the post has been edited the last time
-	AllowsComments bool           `json:"allows_comments"`         // Tells if users can reference this PostID as the parent
-	Subspace       string         `json:"subspace"`                // Identifies the application that has posted the message
-	OptionalData   OptionalData   `json:"optional_data,omitempty"` // Arbitrary data that can be used from the developers
-	Creator        sdk.AccAddress `json:"creator"`                 // Creator of the Post
-	Medias         PostMedias     `json:"medias,omitempty"`        // Contains all the medias that are shared with the post
-	PollData       *PollData      `json:"poll_data,omitempty"`     // Contains the poll details, if existing
+	PostID         PostID         `json:"id" yaml:"id" `                                          // Unique id
+	ParentID       PostID         `json:"parent_id" yaml:"parent_id"`                             // Post of which this one is a comment
+	Message        string         `json:"message" yaml:"message"`                                 // Message contained inside the post
+	Created        time.Time      `json:"created" yaml:"created"`                                 // RFC3339 date at which the post has been created
+	LastEdited     time.Time      `json:"last_edited" yaml:"last_edited"`                         // RFC3339 date at which the post has been edited the last time
+	AllowsComments bool           `json:"allows_comments" yaml:"allows_comments"`                 // Tells if users can reference this PostID as the parent
+	Subspace       string         `json:"subspace" yaml:"subspace"`                               // Identifies the application that has posted the message
+	OptionalData   OptionalData   `json:"optional_data,omitempty" yaml:"optional_data,omitempty"` // Arbitrary data that can be used from the developers
+	Creator        sdk.AccAddress `json:"creator" yaml:"creator"`                                 // Creator of the Post
+	Medias         PostMedias     `json:"medias,omitempty" yaml:"medias,omitempty"`               // Contains all the medias that are shared with the post
+	PollData       *PollData      `json:"poll_data,omitempty" yaml:"poll_data"`                   // Contains the poll details, if existing
 }
 
 func NewPost(id, parentID PostID, message string, allowsComments bool, subspace string,
@@ -209,8 +209,10 @@ func (p Post) Validate() error {
 		return err
 	}
 
-	if err := p.PollData.Validate(); err != nil {
-		return err
+	if p.PollData != nil {
+		if err := p.PollData.Validate(); err != nil {
+			return err
+		}
 	}
 
 	return nil